Backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.4
django-cors-headers==3.1.0
requests
unipath
django-filter
